Debugging and Documentation Left
I've added in all the forms that are set to be in this project, along with all the contents within such as the fields and options to choose from. CSS Stylesheets were also added in, but comments and a couple bug fixes are still needed. Just like yesterday, I plan to finish these later tonight.